Fixes instance_find(), ignores some .c_incls.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@415 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Universal_System/instance.cpp
@@ -57,7 +57,7 @@ bool instance_exists(int obj)
   return enigma::fetch_instance_by_int(obj) != NULL;
 }
 
-int instance_find(double obj,double n)
+int instance_find(int obj,int n)
 {
   int num=(int)n;
   int nth=0;
